Moved Server.js into server folder.
